+++ src/main/java/net/yxiao233/ifeu/common/registry/ModRarity.java
@@ -5,4 +5,5 @@ import net.minecraft.world.item.Rarity;
 
 public class ModRarity {
     public static final Rarity LEGENDARY = Rarity.create("legendary",ChatFormatting.GOLD);
+    public static final Rarity MYTHOLOGICAL = Rarity.create("ultimate",ChatFormatting.DARK_RED);
 }