+++ src/main/java/net/yxiao233/ifeu/common/registry/ModTags.java
@@ -12,6 +12,7 @@ import net.yxiao233.ifeu.IndustrialForegoingExtraUpgrades;
 public class ModTags {
     public static class Blocks{
         public static final TagKey<Block> NEEDS_DRAGON_STAR_TOOL = createTag("need_dragon_star_tool");
+        public static final TagKey<Block> MACHINE_FRAME_ULTIMATE = TagUtil.getBlockTag(new ResourceLocation("ifeu:machine_frame/ultimate"));
         private static TagKey<Block> createTag(String name){
             return BlockTags.create(new ResourceLocation(IndustrialForegoingExtraUpgrades.MODID,name));
         }
@@ -20,6 +21,7 @@ public class ModTags {
         public static final TagKey<Item> GEARS_NETHERITE = TagUtil.getItemTag(new ResourceLocation("forge:gears/netherite"));
         public static final TagKey<Item> GEARS = TagUtil.getItemTag(new ResourceLocation("forge:gears"));
         public static final TagKey<Item> GEARS_SCULK = TagUtil.getItemTag(new ResourceLocation("forge:gears/sculk"));
+        public static final TagKey<Item> MACHINE_FRAME_ULTIMATE = TagUtil.getItemTag(new ResourceLocation("ifeu:machine_frame/ultimate"));
 
         private static TagKey<Item> createTag(String name){
             return ItemTags.create(new ResourceLocation(IndustrialForegoingExtraUpgrades.MODID,name));